+++ qteditor/QtEditor/assetbrowser.cpp
@@ -1,112 +1,22 @@
 #include "assetbrowser.h"
 #include "ui_assetbrowser.h"
-#include <qfilesystemmodel.h>
-#include <qlistwidget.h>
-#include <qmenu.h>
-#include <qprocess.h>
+#include "file_system_watcher.h"
 #include "core/crc32.h"
 #include "core/resource.h"
 #include "core/resource_manager.h"
 #include "editor/editor_client.h"
 #include "editor/editor_server.h"
 #include "engine/engine.h"
-
-#include <Windows.h>
+#include <qfilesystemmodel.h>
+#include <qlistwidget.h>
+#include <qmenu.h>
+#include <qprocess.h>
 
 
 struct ProcessInfo
 {
 	class QProcess* m_process;
 	Lumix::string m_path;
-};
-
-
-// http://qualapps.blogspot.sk/2010/05/understanding-readdirectorychangesw_19.html
-class FileSystemWatcher
-{
-	public:
-		static void wcharToCharArray(const WCHAR* src, char* dest, int len)
-		{
-			for(unsigned int i = 0; i < len / sizeof(WCHAR); ++i)
-			{
-				dest[i] = static_cast<char>(src[i]); 
-			}
-			dest[len / sizeof(WCHAR)] = '\0';
-		}
-
-		void start(LPCWSTR path)
-		{
-			m_overlapped.hEvent = this;
-			m_handle = CreateFile(path, FILE_LIST_DIRECTORY, FILE_SHARE_DELETE | FILE_SHARE_READ | FILE_SHARE_WRITE, NULL, OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS | FILE_FLAG_OVERLAPPED, NULL);
-			ReadDirectoryChangesW(m_handle, m_info, sizeof(m_info), TRUE, FILE_NOTIFY_CHANGE_FILE_NAME | FILE_NOTIFY_CHANGE_LAST_WRITE, &m_received, &m_overlapped, callback);
-		}
-
-		static void CALLBACK callback(DWORD errorCode, DWORD tferred, LPOVERLAPPED over)
-		{
-			ASSERT(errorCode == 0);
-			FileSystemWatcher* watcher = (FileSystemWatcher*)over->hEvent;
-			if(tferred > 0)
-			{
-				FILE_NOTIFY_INFORMATION* info = &watcher->m_info[0];
-				while(info)
-				{
-					switch(info->Action)
-					{
-						case FILE_ACTION_ADDED:
-						case FILE_ACTION_MODIFIED:
-							{
-								char tmp[MAX_PATH];
-								wcharToCharArray(info->FileName, tmp, info->FileNameLength);
-								watcher->m_asset_browser->emitFileChanged(tmp);
-							}
-							break;
-						default:
-							ASSERT(false);
-							break;
-					}
-					info = info->NextEntryOffset == 0 ? NULL : (FILE_NOTIFY_INFORMATION*)(((char*)info) + info->NextEntryOffset);
-				}
-			}
-			BOOL b = ReadDirectoryChangesW(watcher->m_handle, watcher->m_info2, sizeof(watcher->m_info), TRUE, FILE_NOTIFY_CHANGE_LAST_WRITE | FILE_NOTIFY_CHANGE_FILE_NAME, &watcher->m_received, &watcher->m_overlapped, callback2);
-			ASSERT(b);
-		}
-		
-		static void CALLBACK callback2(DWORD errorCode, DWORD tferred, LPOVERLAPPED over)
-		{
-			ASSERT(errorCode == 0);
-			FileSystemWatcher* watcher = (FileSystemWatcher*)over->hEvent;
-			if(tferred > 0)
-			{
-				FILE_NOTIFY_INFORMATION* info = &watcher->m_info2[0];
-				while(info)
-				{
-					switch(info->Action)
-					{
-						case FILE_ACTION_ADDED:
-						case FILE_ACTION_MODIFIED:
-							{
-								char tmp[MAX_PATH];
-								wcharToCharArray(info->FileName, tmp, info->FileNameLength);
-								watcher->m_asset_browser->emitFileChanged(tmp);
-							}
-							break;
-						default:
-							ASSERT(false);
-							break;
-					}
-					info = info->NextEntryOffset == 0 ? NULL : (FILE_NOTIFY_INFORMATION*)(((char*)info) + info->NextEntryOffset);
-				}
-			}
-			BOOL b = ReadDirectoryChangesW(watcher->m_handle, watcher->m_info, sizeof(watcher->m_info), TRUE, FILE_NOTIFY_CHANGE_LAST_WRITE | FILE_NOTIFY_CHANGE_FILE_NAME, &watcher->m_received, &watcher->m_overlapped, callback);
-			ASSERT(b);
-		}
-
-		FILE_NOTIFY_INFORMATION m_info[10];
-		FILE_NOTIFY_INFORMATION m_info2[10];
-		HANDLE m_handle;
-		DWORD m_received;
-		OVERLAPPED m_overlapped;
-		AssetBrowser* m_asset_browser;
 };
 
 
@@ -120,10 +30,9 @@ AssetBrowser::AssetBrowser(QWidget* parent) :
 	QDockWidget(parent),
 	m_ui(new Ui::AssetBrowser)
 {
-	m_watcher = new FileSystemWatcher;
-	m_watcher->start(QDir::currentPath().toStdWString().c_str());
+	m_watcher = FileSystemWatcher::create(QDir::currentPath().toLatin1().data());
+	m_watcher->getCallback().bind<AssetBrowser, &AssetBrowser::onFileSystemWatcherCallback>(this);
 	m_base_path = QDir::currentPath();
-	m_watcher->m_asset_browser = this;
 	m_client = NULL;
 	m_server = NULL;
 	m_ui->setupUi(this);
@@ -147,6 +56,12 @@ AssetBrowser::~AssetBrowser()
 {
 	delete m_ui;
 	delete m_model;
+}
+
+
+void AssetBrowser::onFileSystemWatcherCallback(const char* path)
+{
+	emitFileChanged(path);
 }
