+++ qteditor/QtEditor/file_system_watcher.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "core/delegate.h"
+#include "core/path.h"
+
+
+class FileSystemWatcher
+{
+	public:
+		static FileSystemWatcher* create(const Lumix::Path& path);
+		static void destroy(FileSystemWatcher* watcher); 
+		virtual Lumix::Delegate<void (const char*)>& getCallback() = 0;
+};
+
